Have moved Selenium to the right directory in the project tree and have changed the chrome driver to a new location which may help us in easy update of the project.
